fixed: int changed to float

+++ arrays/bucketsort.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 void insertionSort(vector<float> &bucket){
     for(int i=1;i<bucket.size();i++){
-        int key = bucket[i];
+        float key = bucket[i];
         int j = i-1;
         while(j>=0 && bucket[j] > key){
             bucket[j+1] = bucket[j];
